edits on sigfox.h to test
Also including file to test sigfox module

+++ lib/TrackerFake/gpsFake.cpp
@@ -2,9 +2,9 @@
 #include "TrackerFake.h"
 #include "gpsFake.h"
 
-float gpsClass::lat() { TrackerFakeInstance(gps)->lat(); };
+double gpsClass::lat() { TrackerFakeInstance(gps)->lat(); };
 
-float gpsClass::lng() { TrackerFakeInstance(gps)->lng(); };
+double gpsClass::lng() { TrackerFakeInstance(gps)->lng(); };
 
 float gpsClass::encode(int value) { return TrackerFakeInstance(gps)->encode(value); };
