+++ lib/TrackerFake/gpsFake.h
@@ -3,8 +3,8 @@
 #include "TrackerFake.h"
 
 struct gpsFake_ {
-    virtual float lat() = 0;
-    virtual float lng() = 0;
+    virtual double lat() = 0;
+    virtual double lng() = 0;
     virtual float encode(int);
     virtual int isUpdated();
 };
@@ -12,8 +12,8 @@ struct gpsFake_ {
 class gpsClass {
  public:
   
-  virtual float lat();
-  virtual float lng();
+  virtual double lat();
+  virtual double lng();
 
   virtual float encode(int value);
   virtual int isUpdated();
